[OPIK-6266][FE] refactor: move extractErrorMessage to lib/errors
The helper is used by 10 call sites across environments, traces, datasets, and the generic query error toast — `lib/tags.ts` is no longer the right home. Relocate it to `lib/errors.ts` and update every import.
