+++ apps/opik-frontend/src/api/datasets/useExperimentBatchUpdateMutation.ts
@@ -4,11 +4,8 @@ import { AxiosError } from "axios";
 import api, { EXPERIMENTS_REST_ENDPOINT } from "@/api/api";
 import { Experiment } from "@/types/datasets";
 import { useToast } from "@/ui/use-toast";
-import {
-  TagUpdateFields,
-  buildTagUpdatePayload,
-  extractErrorMessage,
-} from "@/lib/tags";
+import { TagUpdateFields, buildTagUpdatePayload } from "@/lib/tags";
+import { extractErrorMessage } from "@/lib/errors";
 
 type UseExperimentBatchUpdateMutationParams = {
   ids: string[];
