+++ apps/opik-frontend/src/api/traces/useThreadBatchUpdateMutation.ts
@@ -3,11 +3,8 @@ import { AxiosError } from "axios";
 
 import api, { THREADS_KEY, TRACES_REST_ENDPOINT } from "@/api/api";
 import { useToast } from "@/ui/use-toast";
-import {
-  TagUpdateFields,
-  buildTagUpdatePayload,
-  extractErrorMessage,
-} from "@/lib/tags";
+import { TagUpdateFields, buildTagUpdatePayload } from "@/lib/tags";
+import { extractErrorMessage } from "@/lib/errors";
 
 type UseThreadBatchUpdateMutationParams = {
   projectId: string;
